+++ src/repr/adjacency_map/mod.rs
@@ -123,6 +123,7 @@ use {
         FilterVertices,
         GrowingNetwork,
         HasArc,
+        HasEdge,
         InNeighbors,
         Indegree,
         IsComplete,
@@ -596,6 +597,12 @@ impl GrowingNetwork for AdjacencyMap {
 impl HasArc for AdjacencyMap {
     fn has_arc(&self, u: usize, v: usize) -> bool {
         self.arcs.get(&u).is_some_and(|set| set.contains(&v))
+    }
+}
+
+impl HasEdge for AdjacencyMap {
+    fn has_edge(&self, u: usize, v: usize) -> bool {
+        self.has_arc(u, v) && self.has_arc(v, u)
     }
 }